esio: fix some unsigned/signed int comparisons
Investigated three cases and determined they look safe.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/esio/trunk@16537 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ esio/restart-rename.c
@@ -62,10 +62,10 @@ snprintf_realloc(char **str, size_t *size, const char *format, ...)
     va_list ap;
     va_start(ap, format);
     while (   0     <= (retval = vsnprintf(*str, *size, format, ap)) // Error?
-           && *size <  (needed = retval + 1)) {                      // Space?
+           && *size <  (size_t) (needed = retval + 1)) {             // Space?
         va_end(ap);
         *size *= 2;                                                  // Space?
-        if (*size < needed) *size = needed;                          // Space!
+        if (*size < (size_t) needed) *size = needed;                 // Space!
         char *p = realloc(*str, *size);                              // Alloc
         if (p) {
             *str = p;
